# openapi.yml
{
   "components":{
      "schemas":{
...
         "JobApiModel":{
            "properties":{
               "name":{
                  "type":"string",
                  "not":{
                     "type":"null"
                  }
               }
            },
            "type":"object",
            "required":[
               "name"
            ]
         },
...
      },
      "securitySchemes":{
         "bearerAuth":{
            "type":"http",
            "scheme":"bearer",
            "bearerFormat":"JWT"
         }
      }
   },
   "info":{
      "title":"Koleg 👩‍💼",
      "version":"1.0.0",
      "description":"Koleg rest api",
      "license":{
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
   "openapi":"3.0.0",
   "paths":{
...
      "/jobs":{
         "get":{
            "operationId":"JobsController.getAll",
            "parameters":[
               {
                  "in":"query",
                  "name":"filter",
                  "required":false,
                  "schema":{
                     "type":"string"
                  }
               },
               {
                  "in":"query",
                  "name":"page",
                  "required":false,
                  "schema":{
                     "type":"number"
                  }
               },
               {
                  "in":"query",
                  "name":"itemsNumber",
                  "required":false,
                  "schema":{
                     "type":"number"
                  }
               }
            ],
            "responses":{
               "200":{
                  "content":{
                     "application/json":{
                        "schema":{
                           "items":{
                              "$ref":"#/components/schemas/JobApiModel"
                           },
                           "type":"array"
                        }
                     }
                  },
                  "description":"A list of jobs"
               }
            },
            "summary":"Get all",
            "tags":[
               "Jobs"
            ],
            "description":"Query all jobs using filter or not.",
            "security":[
               {
                  "bearerAuth":[

                  ]
               }
            ]
         },
         "post":{
            "operationId":"JobsController.post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/JobApiModel"
                     }
                  }
               },
               "description":"JobApiModel",
               "required":false
            },
            "responses":{
               "200":{
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseModel"
                        }
                     }
                  },
                  "description":"Response model with job id"
               }
            },
            "summary":"Post",
            "tags":[
               "Jobs"
            ],
            "description":"Create new job.",
            "security":[
               {
                  "bearerAuth":[

                  ]
               }
            ]
         }
      },
      "/jobs/{name}":{
         "get":{
            "operationId":"JobsController.get",
            "parameters":[
               {
                  "in":"path",
                  "name":"name",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseModel"
                        }
                     }
                  },
                  "description":"Response of job query."
               }
            },
            "summary":"Get",
            "tags":[
               "Jobs"
            ],
            "description":"Request job using his id.",
            "security":[
               {
                  "bearerAuth":[

                  ]
               }
            ]
         },
         "delete":{
            "operationId":"JobsController.delete",
            "parameters":[
               {
                  "in":"path",
                  "name":"name",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseModel"
                        }
                     }
                  },
                  "description":"Response of delete job query."
               }
            },
            "summary":"Delete",
            "tags":[
               "Jobs"
            ],
            "description":"Delete job using his name.",
            "security":[
               {
                  "bearerAuth":[

                  ]
               }
            ]
         }
      },
...
   },
...
   "servers":[
      {
         "url":"http://localhost:8080/",
         "description":"Local development server"
      },
      {
         "url":"https://api.dev.koleg.tk/",
         "description":"Development server"
      },
      {
         "url":"https://api.koleg.tk/",
         "description":"Production server"
      }
   ]
}